Patch#47: Added Shortcut Keys - Added Shortcut Keys for Dialogs, Transactions, Purchasing - Added more guard for permissions in dialogs - Fix bug in Price Unit and Master Stock - Added Icon for Price Unit

+++ dialogs/master_stock_dialog/services/master_stock_dialog_services.py
@@ -1,10 +1,22 @@
 from dialogs.master_stock_dialog.repositories.master_stock_dialog_repositories import MasterStockDialogRepository
 from dialogs.master_stock_dialog.models.master_stock_dialog_models import MasterStockModel
 
+from generals.constants import (
+    PERM_C_PRODUCTS, PERM_U_PRODUCTS, PERM_D_PRODUCTS,
+)
+from generals.messages import ( 
+    ERR_PERM_C_PRODUCTS, ERR_PERM_U_PRODUCTS, ERR_PERM_D_PRODUCTS,
+)
+from response.response_message import ResponseMessage
+from generals.permission_manager import PermissionManager
+
+
 class MasterStockDialogService:
     def __init__(self):
         self.repository = MasterStockDialogRepository()
-        
+        self.permission_manager = PermissionManager()
+
+
     def get_suppliers(self, search_text: str = ''):
         return self.repository.get_suppliers(search_text)
     
@@ -26,12 +38,22 @@ class MasterStockDialogService:
     
 
     def submit_master_stock(self, master_stock_form_data: MasterStockModel):
+        if not self.permission_manager.has_permission(PERM_C_PRODUCTS):
+            return ResponseMessage.fail(message=ERR_PERM_C_PRODUCTS)
+
         return self.repository.submit_master_stock(master_stock_form_data)
     
 
     def update_master_stock(self, master_stock_form_data: MasterStockModel):
+        if not self.permission_manager.has_permission(PERM_U_PRODUCTS):
+            return ResponseMessage.fail(message=ERR_PERM_U_PRODUCTS)
+
         return self.repository.update_master_stock(master_stock_form_data)
 
 
     def delete_master_stock_by_sku(self, sku: str):
+        if not self.permission_manager.has_permission(PERM_D_PRODUCTS):
+            return ResponseMessage.fail(message=ERR_PERM_D_PRODUCTS)
+
         return self.repository.delete_master_stock_by_sku(sku)
+
